pip_requirements.txt
sorl-thumbnail==3.2.5
python-ldap
simplejson
